✨ [BE] feat: 랭킹 페이지 조회 API 추가

+++ altteul-be/src/main/java/com/c203/altteulbe/ranking/util/TodayRankingPredicate.java
@@ -5,12 +5,15 @@ import static com.c203.altteulbe.user.persistent.entity.QUser.*;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import org.springframework.util.StringUtils;
 
+/*
+ * 동적 쿼리를 위한 조건 빌딩
+ */
 public class TodayRankingPredicate {
 	public static BooleanExpression nicknameContains(String nickname) {
 		return StringUtils.hasText(nickname) ? user.nickname.containsIgnoreCase(nickname) : null;
 	}
 
-	public static BooleanExpression tierEquals(Integer tierId) {
+	public static BooleanExpression tierEquals(Long tierId) {
 		return (tierId != null) ? tier.id.eq(tierId.longValue()) : null;
 	}
